Found a way sorting can work on percentile with real numbers (although FE does not use it, may be important for Scripting API)

+++ graylog2-server/src/main/java/org/graylog/plugins/views/search/searchtypes/pivot/series/Percentile.java
@@ -51,7 +51,7 @@ public abstract class Percentile implements SeriesSpec, HasField {
 
     @Override
     public Optional<String> multiValueAggSubfieldName() {
-        return Optional.of(String.valueOf(percentile().longValue()));
+        return Optional.of(String.valueOf(percentile()));
     }
 
     public abstract Builder toBuilder();
